+++ src/main/kotlin/spring/studies/transactions/trackzilla/service/TicketService.kt
@@ -4,9 +4,9 @@ import spring.studies.transactions.trackzilla.model.Ticket
 
 interface TicketService {
     fun getAllTickets(): List<Ticket>
-    fun getTicketById(ticketId: Int): Ticket?
+    fun getTicketById(ticketId: Long): Ticket?
     fun addTicket(ticket: Ticket)
     fun updateTicket(ticket: Ticket)
-    fun deleteTicket(ticketId: Int)
-    fun closeTicket(ticketId: Int)
+    fun deleteTicket(ticketId: Long)
+    fun closeTicket(ticketId: Long)
 }
